Close unwanted file descriptors.
In scgi_server.py, spawn_child() is called at startup to start the
first child and also from delegate_request() when more children are
needed. In the latter case, the parameter 'conn' is passed to
spawn_child() so that the newly-created child knows to close the
file descriptor it has inherited but doesn't need.

The bug is that in the latter case the new child also inherits
various other file descriptors which are not similarly closed,
namely the Unix sockets to its elder siblings, and the TCP listener
socket.