Create a new server that provides session affinity.
Add a new module session_server.py.  The default server tries to run
a minimal number of worker processes and passes requests to the
workers in the order they have been started.  This new server peeks
at the HTTP session cookie and associates a worker process with each
session.  There are advantages to each approach.