Improve exception handling in Child.process().
Since Python 3.5, EINTR is no longer returned to user code.  So, remove
handling of it.  In Python 3, socket.error is OSError.  So, use OSError
directly rather than socket.error.

If an exception occurs inside Child.process(), close the connection to
the child.  That will cause the child to exit and be reaped.  The
previous behavior of raising an exception from process() could cause the
master SCGI process to exit in some cases.  We want to avoid that.