Limit retries in delegate_request().
If we continuously fail to pass a request to a child, give up after a
certain number of attempts (30).  Otherwise, delegate_request() will
never return.  If this happens, something is seriously wrong with the
SCGI application but it seems slightly better to start dropping incoming
requests in that case rather than never making progress.